Refactor FppTest
