+++ FppTest/CMakeLists.txt
@@ -12,5 +12,6 @@ include("${CMAKE_CURRENT_LIST_DIR}/../cmake/FPrime.cmake")
 include("${CMAKE_CURRENT_LIST_DIR}/../cmake/FPrime-Code.cmake")
 
 if (BUILD_TESTING AND NOT __FPRIME_NO_UT_GEN__)
-    add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/test/")
+    add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/enum/")
+    add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/array/")
 endif()
