+++ FppTest/array/enum.fpp
