+++ FppTest/array/format.fpp
