+++ FppTest/array/struct.fpp
