+++ FppTest/enum/CMakeLists.txt
