+++ FppTest/enum/default.fpp
