+++ FppTest/enum/explicit.fpp
