+++ FppTest/enum/implicit.fpp
