+++ FppTest/enum/interval.fpp
