+++ FppTest/enum/main.cpp
