+++ FppTest/enum/serialize_type.fpp
