--- FppTest/test/CMakeLists.txt
@@ -1,2 +1,0 @@
-add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/enum/")
-add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/array/")
